requirements.txt
alembic==1.0.8
bcrypt==3.1.6
blinker==1.4
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
Click==7.0
dbf-light==0.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.0
mysqlclient==1.4.2.post1
Pillow==5.4.1
psycopg2==2.8.2
pycbrf==0.3.0
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.2.17
urllib3==1.25.3
Werkzeug==0.14.1
WTForms==2.2.1
whoosh==2.7.4
blinker==1.4
flask-msearch==0.2.4